feat(brew): use Raycast instead of Alfred
Raycast is free of use. Alfred has a lot of the useful features, e.g. the workflows, under the paid Powerpack. Raycast has better, more useful feautures in its free version. I've also used the Powerpack, so am familiar with the paid version of Alfred too. For me personally, the free version of Raycast is the better option, at the moment at least.