fix(shottr): only save on Esc, preventing double files
With saving on grap as well as Esc, we end up with two files. This would be useful when editing, to get the plain file if you forget to save it. With this configuration the only case where the file would not be saved is when Shottr quits after grabbing the screenshot and before closing the edit window. That risk is very low, so for me this is the better option.