feat: use `mise` to manage Node.js Rust and Golang versions
I used nvm for Node.js before and nothing for Rust and Golang. One of the biggest issues I had with nvm is the load time, which took at least three full seconds to load. Unfortunately the loading needs to be part of the Zsh file, or you can lazy load when you e.g. call `node`. Mise seemed the better option to me, and loads extremely fast. Additionally, for the global Node.js packages we can make use of the `postinstall` option of mise.